requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
confluent-kafka==1.7.0
Deprecated==1.2.13
Django==3.1.7
django-cors-headers==3.7.0
django-mysql==3.11.1
django-redis==4.12.1
djangorestframework==3.12.2
Faker==6.6.1
idna==3.3
mysqlclient==2.0.3
packaging==21.3
psycopg2==2.9.3
PyJWT==2.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.56.0
text-unidecode==1.3
urllib3==1.26.9
wrapt==1.14.1
